Added auto overrides for verbs, now at 100 percent
It looks like a lot of the problems with the verbs are because of two main
issues: improper handling of -či verbs (quite possibly due to unicode errors in
my create_lexica.py script), and verbs that have two possible stems.  Both of
those could probably be handled in the regular lexicon with some better
processing.  When they are fixed, it's easy to generate a new (hopefully
smaller) auto_overrides file.  For now, though, we'll leave this as it is.