Finished noun auto overrides.
Now at essentially 100 percent recall, and 98.8 percent precision overall.  The
precision errors are due to both nouns and adjectives, mostly proper (and thus
gender-restricted) forms of both of those.

The noun auto-overrides were less obvious than adjectives and verbs; there are
some nouns with two possible forms, and a significant number with irregular
stems (like sin -> sinovi, adding a -t syllable, and other similar things).
The stem changing code might help some here, but a lot of it is just
unpredictable, anyway.

I am calling this analyzer finished.  There are some improvements that could be
made (as mentioned in the last few commit messages), but it seems quite
complete to me.  There are two things left to do: write a script that downloads
a random Slovene news article and runs it through here looking to see how
frequently there are unseen words, and implement an unseen word guesser.
Because of the size of my original corpus, I'm guessing that the unseen word
guesser will mostly only be necessary for proper nouns (and adjectives).  But
the script that grabs a news article will let me quantify that.